Rename reducer/index.js to songReducer
